fusion/requirements.txt
asgiref==3.8.1
coverage==7.6.10
dj-static==0.0.6
Django==5.1.3
django-stdimage==6.0.2
model-bakery==1.20.1
pillow==11.1.0
psycopg==3.2.4
psycopg2-binary==2.9.10
sqlparse==0.5.2
static3==0.7.0
typing_extensions==4.12.2
tzdata==2024.2
